backend_python/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
bcrypt==4.1.3
passlib==1.7.4
python-multipart
email-validator
python-jose[cryptography]
